+++ lib/model/webservice/activity_summary/activity_summary.dart
@@ -7,9 +7,9 @@ class ActivitySummary {
   @JsonKey(name: 'total_distance', required: true, includeIfNull: false)
   double totalDistance;
   @JsonKey(name: 'total_moving_time', required: true, includeIfNull: false)
-  double totalMovingTime;
+  int totalMovingTime;
   @JsonKey(name: 'total_elapsed_time', required: true, includeIfNull: false)
-  double totalElapsedTime;
+  int totalElapsedTime;
   @JsonKey(name: 'total_elevatiion_gain', required: true, includeIfNull: false)
   double totalElevatiionGain;
